+++ src/app/modal.rs
@@ -85,6 +85,7 @@ pub enum SettingItem {
     SpotifyStartOnSpotify,
     SpotifyClientId,
     SpotifyPlaybackMode,
+    SpotifyCrossfade,
     SpotifyRadioMode,
     YoutubeRadioMode,
     YoutubeSponsorblock,
@@ -128,6 +129,7 @@ impl SettingItem {
             Self::SpotifyStartOnSpotify => t("config.setting.spotify_start_on_spotify"),
             Self::SpotifyClientId => t("config.setting.spotify_client_id"),
             Self::SpotifyPlaybackMode => t("config.setting.spotify_playback_mode"),
+            Self::SpotifyCrossfade => t("config.setting.spotify_crossfade"),
             Self::AutoUpdate => t("config.setting.auto_update"),
             Self::DiscordRpc => t("config.setting.discord_rpc"),
             Self::ReplayOnboarding => t("config.setting.replay_onboarding"),
@@ -171,6 +173,7 @@ impl SettingItem {
             Self::SpotifyStartOnSpotify => "config.tooltip.spotify_start_on_spotify",
             Self::SpotifyClientId => "config.tooltip.spotify_client_id",
             Self::SpotifyPlaybackMode => "config.tooltip.spotify_playback_mode",
+            Self::SpotifyCrossfade => "config.tooltip.spotify_crossfade",
             Self::AutoUpdate => "config.tooltip.auto_update",
             Self::DiscordRpc => "config.tooltip.discord_rpc",
             Self::ReplayOnboarding => "config.tooltip.replay_onboarding",
@@ -194,7 +197,8 @@ impl SettingItem {
             | Self::SpotifyStopOnQuit
             | Self::SpotifyStartOnSpotify
             | Self::SpotifyClientId
-            | Self::SpotifyPlaybackMode => "config.group.spotify",
+            | Self::SpotifyPlaybackMode
+            | Self::SpotifyCrossfade => "config.group.spotify",
             Self::YoutubeRadioMode
             | Self::YoutubeSponsorblock
             | Self::YoutubeCrossfade
@@ -253,6 +257,7 @@ pub fn settings_items(duck_enabled: bool, screensaver_active: bool) -> Vec<Setti
         SettingItem::SpotifyStartOnSpotify,
         SettingItem::SpotifyClientId,
         SettingItem::SpotifyPlaybackMode,
+        SettingItem::SpotifyCrossfade,
         SettingItem::YoutubeRadioMode,
         SettingItem::YoutubeSponsorblock,
         SettingItem::YoutubeCrossfade,